# Add multibyte character support to encoder