This PR adds multicharacter byte support to the encoder. 

Resolves #11 

### Benchmarks

```
name                                    old time/op    new time/op    delta
Marshal_MixedData_1-16                    2.76µs ± 3%    2.85µs ± 3%     ~     (p=0.222 n=5+5)
Marshal_MixedData_1000-16                 1.95ms ± 5%    2.15ms ± 5%  +10.58%  (p=0.008 n=5+5)
Marshal_MixedData_100000-16                193ms ± 8%     203ms ± 7%     ~     (p=0.151 n=5+5)
Marshal_String-16                          928ns ± 2%     978ns ± 3%   +5.35%  (p=0.016 n=5+5)
Marshal_StringPtr-16                       962ns ± 3%    1040ns ± 9%   +8.11%  (p=0.008 n=5+5)
Marshal_Int64-16                           920ns ± 1%     980ns ± 5%   +6.57%  (p=0.016 n=4+5)
Marshal_Float64-16                        1.16µs ± 2%    1.24µs ± 4%   +6.62%  (p=0.008 n=5+5)
Marshal_Bool-16                            952ns ± 8%     970ns ± 4%     ~     (p=0.421 n=5+5)

name                                    old alloc/op   new alloc/op   delta
Marshal_MixedData_1-16                    5.00kB ± 0%    5.06kB ± 0%   +1.12%  (p=0.008 n=5+5)
Marshal_MixedData_1000-16                 1.02MB ± 0%    1.08MB ± 0%   +5.88%  (p=0.016 n=4+5)
Marshal_MixedData_100000-16               84.7MB ± 0%    90.7MB ± 0%   +7.08%  (p=0.016 n=5+4)
Marshal_String-16                         4.34kB ± 0%    4.35kB ± 0%   +0.37%  (p=0.008 n=5+5)
Marshal_StringPtr-16                      4.34kB ± 0%    4.37kB ± 0%   +0.74%  (p=0.008 n=5+5)
Marshal_Int64-16                          4.32kB ± 0%    4.34kB ± 0%   +0.37%  (p=0.008 n=5+5)
Marshal_Float64-16                        4.34kB ± 0%    4.38kB ± 0%   +0.74%  (p=0.008 n=5+5)
Marshal_Bool-16                           4.32kB ± 0%    4.34kB ± 0%   +0.37%  (p=0.008 n=5+5)

name                                    old allocs/op  new allocs/op  delta
Marshal_MixedData_1-16                      36.0 ± 0%      24.0 ± 0%  -33.33%  (p=0.008 n=5+5)
Marshal_MixedData_1000-16                  30.0k ± 0%     18.0k ± 0%  -39.99%  (p=0.008 n=5+5)
Marshal_MixedData_100000-16                3.00M ± 0%     1.80M ± 0%  -40.00%  (p=0.016 n=5+4)
Marshal_String-16                           9.00 ± 0%     10.00 ± 0%  +11.11%  (p=0.008 n=5+5)
Marshal_StringPtr-16                        9.00 ± 0%     11.00 ± 0%  +22.22%  (p=0.008 n=5+5)
Marshal_Int64-16                            8.00 ± 0%      9.00 ± 0%  +12.50%  (p=0.008 n=5+5)
Marshal_Float64-16                          11.0 ± 0%      12.0 ± 0%   +9.09%  (p=0.008 n=5+5)
Marshal_Bool-16                             9.00 ± 0%     10.00 ± 0%  +11.11%  (p=0.008 n=5+5)
```